+++ sdk/spring/azure-spring-integration-test/src/main/java/com/azure/spring/integration/test/support/reactor/SendOperationTest.java
@@ -4,12 +4,14 @@
 package com.azure.spring.integration.test.support.reactor;
 
 import com.azure.spring.integration.core.api.reactor.SendOperation;
-import com.google.common.collect.ImmutableMap;
 import org.junit.jupiter.api.Test;
 import org.springframework.core.NestedRuntimeException;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.support.GenericMessage;
 import reactor.core.publisher.Mono;
+
+import java.util.HashMap;
+import java.util.Map;
 
 import static org.assertj.core.api.Fail.fail;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -20,11 +22,17 @@ public abstract class SendOperationTest<O extends SendOperation> {
 
     protected String consumerGroup = "consumer-group";
     protected String destination = "event-hub";
-    protected Message<?> message = new GenericMessage<>("testPayload",
-        ImmutableMap.of("key1", "value1", "key2", "value2"));
+    protected Message<?> message;
     protected Mono<Void> mono = Mono.empty();
     protected String payload = "payload";
     protected O sendOperation;
+
+    public SendOperationTest() {
+        Map<String, Object> valueMap = new HashMap<>(2);
+        valueMap.put("key1", "value1");
+        valueMap.put("key2", "value2");
+        message = new GenericMessage<>("testPayload", valueMap);
+    }
 
     protected abstract void setupError(String errorMessage);