+++ sdk/spring/azure-spring-boot-starter-keyvault-certificates/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 
 ## 3.1.0-beta.1 (Unreleased)
+This release is compatible with Spring Boot 2.5.0 - 2.5.3.
 ### Dependency Upgrades
-- Upgrade to [spring-boot-dependencies:2.5.2](https://repo.maven.apache.org/maven2/org/springframework/boot/spring-boot-dependencies/2.5.2/spring-boot-dependencies-2.5.2.pom).
+- Upgrade to [spring-boot-dependencies:2.5.3](https://repo.maven.apache.org/maven2/org/springframework/boot/spring-boot-dependencies/2.5.3/spring-boot-dependencies-2.5.3.pom).
 
 ## 3.0.1 (2021-07-01)
 ### Bug Fixes